+++ wemake_python_styleguide/visitors/base.py
@@ -117,7 +117,7 @@ class BaseVisitor(metaclass=abc.ABCMeta):
     def add_violation(self, violation: BaseViolation) -> None:
         """Adds violation to the visitor."""
         # It is not allowed to add disabled violations:
-        assert violation.disabled_since is None  # noqa: S101
+        assert violation.disabled_since is None, violation.code  # noqa: S101
         self.violations.append(violation)
 
     @abc.abstractmethod